Move DTO to ApplicationCore project
