--- src/ApplicationCore/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ApplicationCore
-{
-    public class Class1
-    {
-
-    }
-}
